+++ pcapparser/__init__.py
@@ -1,2 +1,2 @@
 __author__ = 'dongliu'
-__version__ = "0.3.2"
+__version__ = "0.3.3"
