+++ pcapparser/textutils.py
